Fix Process module bug - should actually alternate between reading from buffer and waitpid
Summary:
The code did not actually follow the comments which were correct. The read
and wait was supposed to alternate between consuming the contents in the FD's
buffer and waitpid. But the consume call was missing....

Which meant that if the process writing to the pipe was blocked on exiting
because of a filled buffer (64KB on Mac), we would always time out (since we never
unblock the writing process by clearing the buffer)

This fixes that.

Reviewed By: dabek

Differential Revision: D6516371

fbshipit-source-id: 1700612159ad6ba37d04de3f04d9e87c0e04d1e5