Hg base rev supports merge commit state
Summary:
When current commit is a merge commit (most often this occurs when you enter
a on-disk merge conflict state), retrieinvg the base revision only considers
one of the parents. We actually want the max of the merge base of the two
parents because it works better for loading saved states. In particular,
the actual contents of the .hhconfig file will be its contents in the greater
of the two merge bases.

In short, instead of using 'ancestor(master,.)' for the merge base, we
also look at 'ancestor(master,p2())' if p2() exists, and then take the
max of the two.

Since we want a Future that depends on the results of two other Futures,
I also add machinery to do that in Future.Merged

Reviewed By: dabek

Differential Revision: D6541410

fbshipit-source-id: 0513f19900f3e46fa1b067df246376fba7640a41