HHVM: Allow only one active debugger at a time - 1/3
Summary:
We are currently seeing a lot of strange and undefined behavior around multiple debuggers (if users mix hphpd and xdebug/Nuclide at the same time).

Currently, the debugger hook is maintained per-request in thread local storage. It's possible for different debuggers to hook different requests.  Additionally, the PHP VM hook API is not being followed as designed. We have a DebuggerHook::attach() function the debugger calls that returns a bool. Clearly, this is intended to return true on success, and false if another debugger has already attached, however there are a bunch of oddities in the code as it stands today:
  1. hphpd never checks the return code of this, and actually segfaults if its attach doesn't succeed
  2. hphpd is calling attach multiple times for the same request in some cases, and expecting it to be ok. But this actually can mess up the refcount of attached hooks, which can cause us to clear the JIT blacklist at an incorrect time
  3. The attach logic currently gives hphpd super wierd special treatment: if another debugger (xdebug) is attached, it silently detaches it and attaches hphpd. **This is a huge problem for XDebug & Nuclide**. XDebug is not notified that it got detached, so from the Nuclide side, it looks like the debugger just stopped working (no breakpoints, exceptions, evals, etc). **This also causes XDebug to leak an XDebugServer, a socket to the client, and a worker thread that polls the socket**. After this, the state of XDebug is probably undefined until HHVM is restarted.

  I see that #3 was necessary because XDebug was blindly hooking every request, even if there is no debugger client attached. This meant that hphpd could never possibly work if XDebug is also enabled in config.

Now that I'm going to be adding a third debugger, this hook needs to be sane. I am tightening the policy to have a notion of an **active debugger** for the entire HHVM instance. The active debugger will then hook any request(s) it is interested in as it does today, via TLS on each request thread. Any debugger that is not active must never hook a thread now, per the debugger hook contract.

Additionally, attach() and detach() are enforced. A debugger is only attached once, and attempting to attach if another debugger is attached will fail.

Reviewed By: rightparen

Differential Revision: D6156893

fbshipit-source-id: e02c72c3d4421e004a316aa9cd5558ce222f697b