HHVM - Make XDebug play nice with other debuggers - 3/3
Summary:
This diff makes it so that XDebug can honor the new contract with DebuggerHook.  XDebug will set itself as the active debugger if and only if a socket can be established with a remote XDebug client.

This breaks XDebug request init into two phases - the first creates the server and attempts to connect to a client. If this succeeds, we attempt to set XDebug as the active debugger, and attach the debugger hook to the request. The remainder of the initialization phase (Dbgp init) then occurs in the debugger hook callback just as it did before this change.

If another debugger is already attached, we inform the XDebug client that we couldn't attach to the request via a stderr stream message (really the only way XDebug offers us to send such a message given the current protocol) and we don't hook the request.

Reviewed By: rightparen

Differential Revision: D6156892

fbshipit-source-id: c445f577f4d8fd5238982774c0eed4c785636df5