Premultiply kContigTable
Summary:
This table contains the number of objects we will pre-allocate
when splitting a larger free block, or allocating space from
a slab. When using it, we were always multiplying by kSizeIndex2Size[i],
where i is the size class being allocated, so do the multiply
when building the table.

Also move the table to memory-manager.cpp which is the only reference to it.

Reviewed By: alexeyt

Differential Revision: D6248729

fbshipit-source-id: 7cf253c66aa15ae9a42be4f7c911b20cf4fe45c6