Runtime enforcement for immutable objects and properties.
Summary: Throw when trying to write a property declared immutable after the object is done being constructed, when trying to bind a property declared immutable, or when trying to write a dynamic property of a class that forbids dynamic properties. (Nothing sets/clears Attribute::IsBeingConstructed yet, so until the compiler changes happen this is actually forbidding all mutation of immutable properties.) This should be the final set of changes in code paths shared between mutable and immutable objects, so perf for ordinary objects should be unaffected by the remaining work.

Reviewed By: jano

Differential Revision: D6396000

fbshipit-source-id: 167b9c31618f7ed23be128e650f78d02d4dae744