Elide some reffiness checks when WarnOnCallByRefAnnotationMismatch is enabled
Summary: Turning on WarnOnCallByRefAnnotationMismatch currently causes a very clear regression, which is likely in large part because we need to guard all FPass instructions on reffiness. This diff updates hhvm to ignore reffiness of FPassC and FPassVNop when the FPassHint is Any, and changes hhbbc to use the Any hint whenever it can prove that the annotations match.

Reviewed By: markw65

Differential Revision: D6542891

fbshipit-source-id: 9c2e5a3a495c9994e715ad926a2a7d5b6f080824