Fix dev-asan mode crashes
Summary:
We have a function to return the approximate top of stack, which
returned a pointer to a local. With detect_stack_use_after_return
enabled, asan builds got an incorrect address, causing fatals and
crashes (the crashes happened because we fatalled due to "stack
overflow" while parsing systemlib).

This uses our DECLARE_FRAME_POINTER macro instead - which doesn't
return exactly the same thing, but I audited all the call sites, and
it seems fine.

Reviewed By: ricklavoie

Differential Revision: D6556548

fbshipit-source-id: 00182db0a2f56d70bd0f0f9d4c8dab1246f35e55