Replace crossing map with object-start bitmask
Summary:
At the start of GC, we initialize data structures used for finding
the beginnings of objects quickly, while filtering pointers. The
the current data structure (crossing map) works well for this purpose,
but serves no other purpose, and is complicated to update incrementally.

A bitmask marking the location of each object takes more space (16k
per slab instead of 8k per slab), but is faster during filtering,
far simpler to update incrementally, and accellerates the sweep phase,
since we can scan the bitmask instead of iterating over the heap.
It's much faster to find the next set bit than to calculate each object's
size to find the next object.

The crossing map had a tunable parameter called LineSize. The bitmask
does not, so the template parameter was removed. That each word
of the bitmask contains 64 bits, and each bit represents 16 bytes,
so the per-slab overhead is (2M/16/8 = 16k).

Note that the bitmap is addressable with different machine-word sizes,
allowing us to consider 8, 16, 32, or 64 bits as a "line", representing
128, 256, 512, or 1024 bytes of memory.

Reviewed By: alexeyt

Differential Revision: D6485873

fbshipit-source-id: d0cd6c52e10cab1be6516c8432bd30144fda34ec