Informant shouldnt reuse old XDB results
Summary:
XDB results are updated live, so we shouldn't reuse old results.
In particular, an old query could contain no result, or a distant
result, but XDB is updated later with a better result. If we keep
reusing the old query result, we won't see the newer better result.

This fixes that.

Reviewed By: dabek

Differential Revision: D6562345

fbshipit-source-id: 8fda0e5b5bd239031ffd259c74d798ec8048aba6