Fix crash when stored ctor is null.
Summary: A crash was reported when attempting to instantiate an abstract final class.  We were not setting the stored constructor (m_ctor) to a valid Func since D5323418.  The issue is in the interpreter we find the constructor function, and look at its metadata prior to raising errors for instantiating an abstract class.  Instead of fixing this in fpushCtorImpl, I went with assigning the default nullCtor to abstract final classes.  This would fix other situations that depend on the m_ctor pointer being non null (I am unsure if there are any, but this is conservative).  There is no additional cost of adding this function pointer since D6320951 made all default nullCtor's reference a single Func.

Reviewed By: markw65

Differential Revision: D6557762

fbshipit-source-id: cadfa7cc8ffcf5a4ed728e71208bed2a9edce43b