fix attribute parsing with colons in legacy parser and FFP
Summary: The legacy parser had an unfortunate bug that prevented it from parsing attributes with colons in XHP tag literals, but these are totally legal attribute names and it correctly parses them in `attribute` declarations. This led to incorrect inference (it just dropped the stuff after the colon) and a parsing error.

Differential Revision: D6570229

fbshipit-source-id: 91af9ea6f0d99f8106206fb664b642ece4061a2d