Adds option for faster, estimated PGO counters.
Summary:
Benchmarks with small amounts of code generate a lot of contention for that code's PGO counters. Using a `decqm` instead of a `decqmlock` removes the lock, easing the contentions but also leading to accuracy issues. This new switch (`Eval.JitPGORacyProfiling`) changes the counters to use `decqm` if accuracy isn't required.
Closes https://github.com/facebook/hhvm/pull/7984

Reviewed By: paulbiss

Differential Revision: D5794197

Pulled By: ottoni

fbshipit-source-id: ba54a0cc76c8405f2894916774b211a42c52b07c