Call reportJitMaturity() after emitting prologue and anchor translations
Summary:
We force JIT maturity to stay at 99 until code.main reaches its max usage.  But
if the translation that pushes code.main over that limit was a prologue or an
anchor, we weren't calling reportJitMaturity() and, since no other regular
translation would be emitted after that, we'd never call reportJitMaturity()
again.  As a result, JIT maturity would never reach 100.

Reviewed By: binliu19

Differential Revision: D6587903

fbshipit-source-id: de6a8dfc13a2d6f4bb3bc44bc39fb42d5f2132e3