do the right thing for loopy programs
Summary:
Semdiff previously had a known unsoundness, in that it would
report an "empty" loop (one containing only control-flow and some
non-stack-altering instructions) as being equivalent to anything. This
is perhaps not terribly likely to show up as a failure mode for hackc,
but is still frightfully embarrassing in a tool with aspirations to
soundness. This fixes that by revising the treatment of one-sided
moves and detecting such loops. There are also a couple of other minor
cleanups that have been gathered along the way.

(Optional remark for those interested: I previously tried to fix this
by a couple of more "sophisticated" methods (step-indices and tracking
goal dependency traces). But I couldn't see how to make either of them
both correct and terminating, so backed off to this rather more vulgar
solution, which suffices in the context of the rest of the analysis.)

Reviewed By: andrewjkennedy

Differential Revision: D6468489

fbshipit-source-id: 7e5732c1b257ee662c32d8976d3c8c41f03a07f9