fix null-coalesce on optional shape fields on shapes under Tabstract
Summary: The flag to determine if we are under the LHS of a null coalesce operator was not properly carried through when iterating over the concrete supertypes in `Tabstract`. Being on the LHS of a `??` operator actually silences some errors with optional and nullable shape fields, so it's important to carry it through. See the new test case for an example.

Reviewed By: kmeht

Differential Revision: D6585935

fbshipit-source-id: 977b1d05e67b28daa4f22989432c06f34a6d1cd8