Peephole for CGetL2; Concat; SetL -> SetOpL
Summary:
In the original sequence, the refcount on the string is always 2, but
we're just about to throw away one of them. A CGetL would have been
transformed into a PushL which also avoids the extra ref; but we don't
have a corresponding instruction for CGetL2. But we don't need it
because we can do SetOpL instead.

Reviewed By: ricklavoie

Differential Revision: D6590094

fbshipit-source-id: 172c51d146967843a01a443e145e8c3459c50bd5