"incremental_errors" option in local config to partially disable "D6393223: Do not recheck files only because they had errors before" if necessary
Summary:
While I still believe in correctness of {D6393223} in general, it seems that it accidentally uncovers and amplifies some other preexisting errors. We took the risk once, and had to revert the release due to a bug that we fix in {D6594459}.

Here I add a flag to temporarily keep the "sweep false positives under the rug" behavior to unblock the release and test the new thing separately.

Note that it's not completely equivalent to reverting D6393223 stack (it's hard to put this entire change behind a flag, which is why it wasn't in the first place) - it just brings part of previous functionality that turns out to be useful back (for now, since it's still a performance killer).

Reviewed By: jamesjwu

Differential Revision: D6596723

fbshipit-source-id: 2a146475ea0f03e858a492e3b79c9da355bbb8c9