Fix a race condition between parsing and decl phases accessing a changing file
Summary: When a file changes quickly on disk, to avoid different phases of typechecker seeing different contents, we want (during the duration of recheck) to either keep it's Full AST in ParserHeap, or keep it's contents in FileHeap. We are missing one case when the file becomes empty (missing / mangled / non-Hack) - it doesn't have AST anymore, so we don't put it in ParserHeap, but we should put it in FileHeap instead.

Reviewed By: jamesjwu

Differential Revision: D6594459

fbshipit-source-id: 121c61adc3244be0a89421ad5ccea4d4aeec8578