Remove special handling of collections in ArrayIter.
Summary:
Remove logic for special handling of collections; instead just iterate the underlying ArrayData.
This introduces one major behavior change: mutating the collection while iterating it will no longer trigger an exception, and iteration will continue over the values that were in the collection pre-mutation (same as when iterating arrays).
This also introduces a more subtle behavior change in one edge case (as seen in the modified test expectation): previously, while iterating a collection, writing to an existing key that had not yet been iterated over would cause the new value to be seen when the loop reached that key. Now, no mutations made while iterating are visible to the iterator at all. The previous behavior in this edge case was inconsistent with the general case (where exceptions would be thrown). We will need to move to this new behavior if we want to shift the implementation of collections from C++ into Hack, but I don't know if/how we can prove that no Hack code is relying on this quirk.

Reviewed By: jano

Differential Revision: D6589319

fbshipit-source-id: e8bd0d33682f5c2f1c2ee4eefcfa59ff55955dd6