disallow declaring XHP attributes as nullable and required
Summary: This is actually not allowed. When we validate attributes, we don't differentiate between null and missing -- null *means* that the attribute should be treated as if it wasn't there and if available, it means to use a default. This should not be declarable because it makes no sense with XHP's type system. This may be changed going forward as we differentiate between nullable and optional attributes, but this is a good step forward.

Reviewed By: usikder

Differential Revision: D6595326

fbshipit-source-id: aabee05baa7f9c5a23fbab8261a0651f35496461