Add Aast module for AnnotatedAST functor
Summary: Currently, the functor which produces both the NAST and TAST is located in the Nast module. This is a bit odd--for organizational purposes it would make sense to move the functor to its own module that both Nast and Tast depend on. This also makes visitor generation easier in future diffs.

Reviewed By: andrewjkennedy

Differential Revision: D6481755

fbshipit-source-id: a0237efc777a06d9ced48768fb09098d7f2fa8e4