Adjust inlining cost limit based on profile counters
Summary:
This diff adjusts the inlining cost limit in RetranslateAll mode.  The average
profile count for the function entries at the time RetranslateAll is triggered
is used as a base.  When the caller's profile count is significantly above that
base value, we now allow inlining of larger callee functions.  This diff also
reduces the inlining cost limit for colder caller functions.

Reviewed By: paulbiss

Differential Revision: D6604727

fbshipit-source-id: b77400bf626d110f03c65ccea8404b9031eeae14