Fix tier_overrides_tier.php
Summary:
This test has been getting more and more flaky recently, and I just
noticed that its because it fails 100% of the time on kernel
4.11.3-27_fbk7_2638_g404b383, in dbgo builds.

I grabbed a devserver, upgraded to 4.11.3-27_fbk7_2638_g404b383, and
it reprod. The problem turns out to be that it sets an RSS limit of
200M, which isn't quite enough in dbgo builds. Its not quite clear why
the kernel matters, but presumably its either enforcing the limits
more strictly, or computing them differently.

In any case, doubling the limit fixes it.

Reviewed By: alexeyt

Differential Revision: D6621892

fbshipit-source-id: efb3b1c42f93068681c7d47b0eeff1db23ab1ac2