Send IDE_IDLE only when needed
Summary:
The design of IDE_IDLE is that it should be sent to the server whenever the IDE is idle (i.e. has finished processing its outstanding queue of messages). If the server tried to respond to IDE messages and also other command-line commands, there were some subtle races. (See serverClientProvider.ml, sleep_and_check).

More precisely, if the server handles any message over the persistent connection, then it will decline to handle command-line commands until it eventually gets an IDE_IDLE. (It doesn't matter if it gets surprlus IDE_IDLE messages).

Inside clientLsp, it would be a pain to keep track of every single message that got send over the persistent connection, so clientLsp used a simple rule: "if clientLsp gets ANY non-tick event, then maybe it caused something to be sent over the persistent connection, so maybe the server will be waiting for IDE_IDLE; let's therefore send an IDE_IDLE on our next tick event. It doesn't matter if we overcompensate and send IDE_IDLE even when it wasn't needed." (Note that Tick events arise only when there are no messages for clientLsp to handle either from client or from server).

The implementation of that simple rule used to be as follows:
```
if event <> Tick then begin
  match \!state with
  | Main_loop menv -> state := Main_loop { menv with needs_idle = true; }
  | _ -> ()
end;
```

But some refactoring turned it into this:
```
match state, event with
| Main_loop menv, Tick -> Main_loop { menv with needs_idle = true; }
| _ -> state
```

Obviously the refactoring was incorrect and changed the sense of it completely, by saying that IDE_IDLE must be sent "as a consequence any Tick event" rather than "as a consequence of any NON-TICK event".

This diff fixes it up.

Reviewed By: dabek

Differential Revision: D6626420

fbshipit-source-id: 2a55a87e5838fb5d2a4eba36de00a1bea5ffa57b