More precise locating of HHVM errors in foreach
Summary:
Currently, errors that relate to the expression `e` in `foreach (e) {b}` are reported at the line of the closing brace. This diff improves the accuracy of HHVM's error reporting so that the line containing `e` is reported instead.

A future diff will improve reporting further so that the range of lines is recorded; currently, it is the *last* line of the range that is shown.

Reviewed By: paulbiss

Differential Revision: D6615883

fbshipit-source-id: ade6768743e43a37eb0d070562e445a568e02408