Stop ignoring systemlib classes during parsing
Summary:
Its wrong, because extensions can be enabled/disabled at runtime, so
the list can change; and because units that try to define systemlib
classes should fatal, rather than silently fail to do so.

This is a hangover from early hphpc days when we tried to produce
working code "no matter what", because we didn't want an error in an
unused unit to prevent a successful compile. We realized a long time
ago that that was the wrong way to go about it, but this particular
case was never noticed/fixed.

The runtime already handled this correctly, and the only problem in
hhbbc was that it needed to fail to resolve user classes when there's
a systemlib class of the same name (otherwise the systemlib class gets
marked non-unique, leading to assertions later on).

Reviewed By: alexeyt

Differential Revision: D6620538

fbshipit-source-id: 35ccc251383f54b46efe421523f259d664598dd7