Fix a bug in inc_pass_sig
Summary:
While comparing what D6585787 does, with what rcgraph_opts does, I
noticed that rcgraph_opts would push an Inc through a Sig, even though
there was a following Req which was part of the same instruction as
the Sig, and which blocked the Inc.

This happens because we create the Sig before dealing with any effects
that only happen on the next branch, and because inc_pass_sig isn't
aware of instruction boundaries, so it thinks it can insert the Inc
right after the taken branch, but before any other effects take place.

This is the simplest fix I could find, and I'm reluctant to do much
more because once I can prove that D6585787 does everything that
rcgraph_opts does (and I'm very close), I can kill rcgraph_opts
altogether.

Reviewed By: edwinsmith

Differential Revision: D6621150

fbshipit-source-id: 250d7b341f710b5046fb04b69563f4062f9fc2e6