Add a String::asCell helper, use it.
Summary: When converting code from o_set to setProp, I noticed that I had to write make_tv<KindOfString>(string.get()) a lot, and this was long and frequently caused me to have to break statements across lines. Make a String::asCell helper that encapsulates this operation, and use it. All this does is make code slightly more compact.

Reviewed By: markw65

Differential Revision: D6637659

fbshipit-source-id: 63a9c6e08c5cdb259b13f028771e73ea6126adc2