Fix HHVM not hitting certain breakpoints
Summary: HHVM is not hitting breakpoints put on lines with multiple opcodes if the first opcode of the line is not executed by the interpreter. This is because the PC of the first opcode in the line is added to the line breakpoints PC filter, but the subsequent opcodes are not. We'll therefore only hit the breakpoint if the PC lands on the first opcode in the line, and miss it if we jump elsewhere in the range.

Reviewed By: rightparen

Differential Revision: D6455132

fbshipit-source-id: ac520763fd5b90f0b6ad1a2d81957ec7a20c786a