HHVM VSDebugExtension - Add DebuggerSession, dummy thread and command queue - 4/N
Summary:
This change adds a DebuggerSession object, which will represent any state tied to the lifetime of a particular debugging session. This is created when a debugger client attaches, and destroyed when it detaches.

Currently the only state the session maintains is starting and running the "dummy" thread. The dummy is a concept that already exists in both hphpd and XDebug. It is an empty PHP request that loads a predetermined startup document (a PHP script in a known location in the WWW repo) that pre-loads symbols, functions, types, whatever.  Any debugger request from the client that is not directed to a particular request thread is invoked in the context of the dummy request. This includes things like issuing evaluation statements at the console REPL while the program is running (as opposed to issuing them when a request is at a breakpoint, in which case the statements are evaluated by the thread that is running the broken-in request).

The path to the startup document for hphpd and XDebug are stored in the server's config file. For this debugger, I am not hard-coding it anywhere on the back end. The debugger client will send  the path along with the VS Code debug Initialize message, which I have not implemented yet.

The other significant thing in this diff is the introduction of command_queue. This is intended to be an efficient mechanism for sending commands that originate at the debugger client to the correct backend thread for processing. Many of the VS Code commands are going to need to execute on the thread for a particular request: set breakpoint needs to set in each thread since classes are loaded into TLS. Commands for getting a variable need to get it from the correct scope, etc.  Additionally, when the program is broken in, request threads need to wrangle up and then "park" and not return to the PHP VM until the program is resumed. Ideally we have a super efficient way to do this (there's at least one place in hphpd that busy-waits, which burns a core :( ). Threads will call into CommandQueue::processCommands which will block them on a condition variable, allowing the thread to be parked in the kernel until a resume command is issued.

Reviewed By: rightparen

Differential Revision: D6187000

fbshipit-source-id: 4006e6367c4e30b2b9c031b0dc12219ed0f5e8f4