HHVM VSDEbugExtension - Skeleton for debugger hook and attach logic - 5/N
Summary: This change adds a skeleton for a PHP debugger hook for the new debug extension. The methods of the hook are not implemented yet. It also now attaches to and detaches from incoming requests when the debugger is enabled, and adds book keeping where we can maintain state of the program (running vs paused) and a map of the request threads the debugger knows about.

Reviewed By: rightparen

Differential Revision: D6217512

fbshipit-source-id: b7afb4f5dc96bb35c7f96b7972e1b4fad4f92b8b