HHVM VSDebugExtension - Add transport polling threads - 6/N
Summary:
This change cleans up the transport to use fd instead of FILE* (thinking ahead to needing to accomodate a TCP-based transport  for attach) and adds a dedicated polling thread for reading and writing from the fd.

The read thread now owns the input buffer, relieving us of the need for locking around using and resizing it. It is the only thread that reads from fd, processes messages and hands them off to Debugger.

The write thread allows us to use a locked queue of outgoing messages so that other writers with messages for the client don't need to block holding a lock while waiting to write out to the fd. The write thread picks messages out of the queue and writes them into the pipe.

Reviewed By: rightparen

Differential Revision: D6266352

fbshipit-source-id: 6d00517e8ed00685169fce414ec24d0d70ae7e10