HHVM VSDEbugExtension - Add socket transport - 7/N
Summary: This diff adds a TCP-socket based transport to allow attaching the debugger to an instance of HHVM running in Server mode. By default, this transport will listen on port 8999 but this can be set in config.hdf. This transport allows at most one debugger client to connect at any given time.

Reviewed By: rightparen

Differential Revision: D6266435

fbshipit-source-id: ac3b9a61da55c598bfacab9b975f5063dfd8f7ec