HHVM VSDebugExtension - Implement debugger command infrastructure - 8/N
Summary:
This diff adds the machinery necessary for receiving VS Code debug protocol commands from the client via one of the debugger transports, and processing them. Currently no commands are implemented yet (we just log an error and issue a response).

Each command will have a target, depending on what context it should be evaluated in:
 - None: commands that need to be executed inline as they are received (think Set Breakpoint, Disconnect, ...)
 - Request: commands are sent to a particular request (think EvaluateOnCallFrame). These are enqueued into the correct thread's CommandQueue to be picked up next time the request calls into the debugger hook
 - Dummy: commands are sent to the dummy request thread (think RuntimeEvaluate)

The debugger will wrap each command execution to ensure an appropriate protocol response or error is sent back to the client.

Reviewed By: rightparen

Differential Revision: D6266497

fbshipit-source-id: b1b76e04a41836ebedd8ff9ee87f26d09df18142