Nuclide HHVM Debugger - Support "run in terminal" - 9/N
Summary:
In order to support running a script in the terminal from the front-end, we need to have a way to launch the script directly on the command line, but have it wait to begin execution until the debugger has had a chance to attach.  Additionally, even though this is "launch script" mode (as opposed to "server" mode), we want to use the TCP socket transport and allow the debugger to attach that way. This is necessary because the script cannot be launched via the wrapper, otherwise STDIN and STDOUT on the terminal would be VS Code JSON messages, which is not what our user would be expecting to see!

This diff adds an additional command line arg where the invoker can specify a port on which to listen for a debugger. This combined with --mode vsdebug will cause the debugger to listen on the TCP port for a debugger client **and hold script startup until a client attaches**. After that, the rest of the flow is normal. If the client detaches / reattaches, the script will continue running.

Reviewed By: rightparen

Differential Revision: D6287668

fbshipit-source-id: 1f335da96afbf1cc1e617776c750b785012f3b64