Nuclide HHVM Debugger - Implement initialize flow - 11/N
Summary:
This diff implements the InitialzeRequest, LaunchRequest and AttachRequest commands sent by the front end client for the VS debug protocol. InitializeRequest sends client capabilities, which for now are just parsed and stored in the debugger session to be used by other commands that need to consider them.  Launch and Attach are handled by the same command on the back end - by this point the wrapper has already launched or attached the script, we need to just take the startup document path (if any) from the message and use it to start the dummy request thread.

Finally, the InitializedEvent is sent to the front end. This signals we are ready to receive breakpoints.

This change also fixes a deadlock shutting down the dummy request thread by ensuring the active session is not deleted while holding the debugger lock. The session destructor must block until the dummy request thread joins (for cleanliness and also because AsyncFunc's destructor asserts otherwise), but shutdown of the hphp request context from the dummy thread will actually try to call back into the extension's requestShutdown routine, which needs to acquire the debugger lock.

Reviewed By: rightparen

Differential Revision: D6318848

fbshipit-source-id: 312455604fb67047ff12642a07151e1d41e997b3