Nuclide HHVM Debugger - Implement LoaderBreak and StopEvent - 13/N
Summary:
This diff implements the LoaderBreak, where after connecting a client we set the program state to paused and block all threads from starting. Stopped events are sent to the front end to cause it to update the thread state.

Note: Not implemented yet in this diff is breaking already-running requests the next time they call into the debug hook. This will be done along with breakpoint implementation in a future diff.

Reviewed By: rightparen

Differential Revision: D6342215

fbshipit-source-id: 225c9cc5d4ba72458d585462df044febea9c4ee0