Nuclide HHVM Debugger - Implement ContinueRequest, ConfigurationDone and init flow - 14/N
Summary:
Adds a ContinueRequest implementation to resume the target. The command thread is responsible for updating m_state in the debugger, whichallows new requests to flow through and execute. Additionally, it sends a ContinueCommand to every paused request's command queue, causing the thread to wake up and resume execution.

This also implements the final init flow with Nuclide, where we have a configuration done event to indicate the client is fully initialized and it's okay to resume the target.

Reviewed By: rightparen

Differential Revision: D6355503

fbshipit-source-id: 1f29a476f42d140644477948e9dcca3e4f36a90e