Strip nullable off expected type before decomposing
Summary:
During type inference, if the expected type is of the form `?t`, Hack currently doesn't
make use of it. We fix this by stripping the nullable off the expected type first. This
enables a few more contexts e.g. when a lambda is passed to a function that expects a
nullable function type.

No new tests, but some error messages change.

Differential Revision: D6655732

fbshipit-source-id: 168b06d4c81392d903f740603cca2ef93cff29fa