Nuclide HHVM Debugger - Support breakpoints - add bp Mgr - 16.1/N
Summary:
This diff adds some prereqs for supporting breakpoints. It adds a breakpoint manager, which maintains a list of breakpoints and breakpoint IDs from the client. Breakpoints are unique to a client session, and as such are stored in the session object and cleared if the client disconnects. The breakpoint manager keeps track of which breakpoints have been resolved by at least 1 request thread, and is responsible for sending breakpoint changed notification events to the client.

A breakpoint has a line location set by the client, but also a resolved location which indicates where HHVM installed the breakpoint in a compilation unit.

The VS protocol allows for column indicies to be specified with breakpoints, but they are currently ignored because the API that the PHP VM gives us only takes in line numbers.

Stubs are in place to handle conditional breakpoints, but that is not implemented yet in this diff.

Reviewed By: rightparen

Differential Revision: D6442210

fbshipit-source-id: 83475f65d721abdb5c835bab3ccec1ad9daeb512