Nuclide HHVM Debugger - Breakpoints: stop reasons and pause/resume synchronization - 16.2/N
Summary:
This diff fixes up the stop reasons sent with the StoppedEvent so that it will be able to indicate an actual reason ("hit breakpoint X") to the frontend rather than just saying "paused" all the time. It also adds logic that will be used to wrangle threads and synchronize the program state when stopping and resuming the target.

We need to be able to handle several pretty complex race cases where request threads are running, pausing or resuming, and the client is issuing commands.

For example, if the client issues a resume command, we set the program state to "Running", but asynchronously wake all the request threads from their command queues. During this time window, a request that resumed first could hit a breakpoint and pause the target again. Similarly, two request threads could each hit a breakpoint and roughly the same time. The debugger client needs to see a serialized linear ordering of events.

No thread may break the program for any reason while a previous Continue command has been issued but has not yet woken all waiting threads. It must wait for the continue to complete before breaking.
No thread may issue a breakpoint event if the program is already stopped due to another breakpoint, exception, or asnyc break. It must wait for that breakpoint to RESUME before it can issue the next one.

Reviewed By: rightparen

Differential Revision: D6442701

fbshipit-source-id: 5599c43acdef8ef168ad314bff558a0780a9e192