Nuclide HHVM Debugger - Breakpoints: Support setting breakpoints from client - 16.3/N
Summary:
This diff adds command support for receving a SetBreakpoints request from the client and storing the breakpint in the breakpint manager, and in individual request's pending breakpoint queues.

Each request maintains 3 sets of breakpoints, since a bp needs to be installed on each request thread in order to hit.
1. Pending breakpoints - these were received from the client but the request thread hasn't done anything with them yet.
2. Unresolved breakpoints - breakpoints were removed from the pending list by the request, but no location in a loaded compilation unit has been resolved yet
3. Resolved breakpoints - these are installed and ready to hit.

Reviewed By: rightparen

Differential Revision: D6443095

fbshipit-source-id: fb78328a094b5767993456228fd68f1d553b65f9