Nuclide HHVM Debugger - Kill launched request on debugger detach - 17/N
Summary:
In launch mode, the target is expected to terminate when the debugger detaches (in our case, this is when the NodeJS wrapper exits). If the script does not exit, the NodeJS wrapper tries to terminate HHVM when the parent NodeJS process exits - but HHVM ignores the signal and keeps running. This leaves the user with a hidden HHVM process running that is no longer visible from Nuclide. The only way to kill it is `kill -9`.

This diff detects when the debugger has exited in launch mode and kills the request on the next PHP opcode by raising a fatal exception in the request thread. This causes the request to exit, and HHVM to shutdown cleanly.

Reviewed By: rightparen

Differential Revision: D6450296

fbshipit-source-id: 159c9d1b56a740414f24a3e9e91f6ab994553efc