Added declarations for mutability
Summary:
This adds all of the backend support for mutability. It does not change or add any typechecking rules for mutable types, but prepares us for a diff that will.

There's currently no sanity checking to make sure <<__Mutable>> only appears in places where it makes sense. I'll check that as a separate pass later.

Reviewed By: KendallHopkins

Differential Revision: D6605935

fbshipit-source-id: 02f35691b2d8de39ebffad7a99f9a497592ee908